+++ vue/components/ui/atoms/dropdown/dropdown.vue
@@ -17,7 +17,7 @@
                     v-for="(item, index) in items.filter(v => v !== null && v !== undefined)"
                     v-else
                     v-bind:key="item.value"
-                    v-on:click="() => click(item, index)"
+                    v-on:click="event => click(event, item, index)"
                     v-on:mouseenter="() => onMouseenter(index, item)"
                     v-on:mouseleave="() => onMouseleave(index)"
                 >
@@ -143,10 +143,10 @@
     width: auto;
 }
 
-.dropdown-container .dropdown > .dropdown-item.selected,
-.dropdown-container .dropdown > .dropdown-item:not(.disabled):hover,
-.dropdown-container .dropdown > .dropdown-item:not(.disabled):active,
-.dropdown-container .dropdown > .dropdown-item:not(.disabled).highlighted {
+.dropdown-container .dropdown > .dropdown-item.highlightable.selected,
+.dropdown-container .dropdown > .dropdown-item.highlightable:not(.disabled):hover,
+.dropdown-container .dropdown > .dropdown-item.highlightable:not(.disabled):active,
+.dropdown-container .dropdown > .dropdown-item.highlightable:not(.disabled).highlighted {
     background-color: $soft-blue;
     font-weight: 700;
 }
@@ -204,6 +204,10 @@ export const Dropdown = {
             default: () => ({})
         },
         visible: {
+            type: Boolean,
+            default: true
+        },
+        highlightable: {
             type: Boolean,
             default: true
         },
@@ -328,7 +332,7 @@ export const Dropdown = {
         if (this.onHideGlobal) this.$bus.$off("hide-global", this.onHideGlobal);
     },
     methods: {
-        click(item, index) {
+        click(event, item, index) {
             if (this.managed) {
                 // invalidates all of the selected data (only one item can
                 // be selected at a time) and then updates the currently
@@ -338,7 +342,17 @@ export const Dropdown = {
                 });
                 this.$set(this.selectedData, index, true);
             }
+
+            // legacy event for backwards compatibility, this shouldn't
+            // be updated as it's missing the "event" as the first argument
             this.$emit("item-clicked", item, index);
+
+            // triggers the click item event with the proper event argument
+            // ready to be manipulated by any listener
+            this.$emit("click:item", event, item, index);
+
+            // hides the dropdown as an item has been clicked, this is considered
+            // to be the default and expected behaviour
             this.hide();
         },
         highlight(index) {
@@ -393,8 +407,9 @@ export const Dropdown = {
             return {
                 separator: item.separator,
                 icon: Boolean(item.icon),
-                highlighted: this.highlightedData[index] || item.highlighted,
-                selected: this.selectedData[index] || item.selected,
+                highlightable: this.highlightable,
+                highlighted: this.highlightable && (this.highlightedData[index] || item.highlighted),
+                selected: this.highlightable && (this.selectedData[index] || item.selected),
                 disabled: this.disabledData[index] || item.disabled
             };
         }